requirements.txt
asgiref==3.4.1
attrs==21.2.0
black==21.8b0
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
Django==3.2.6
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
Faker==8.12.1
flake8==3.9.2
gunicorn==20.1.0
idna==3.2
inflection==0.5.1
iniconfig==1.1.1
isort==5.9.3
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
mccabe==0.6.1
mixer==7.1.2
mypy-extensions==0.4.3
openapi-codec==1.3.2
packaging==21.0
pathspec==0.9.0
Pillow==8.3.1
platformdirs==2.2.0
pluggy==0.13.1
psycopg2-binary==2.9.1
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.4
pytest-cov==2.12.1
pytest-django==4.4.0
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
regex==2021.8.28
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
simplejson==3.17.5
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
textdistance==4.2.1
toml==0.10.2
tomli==1.2.1
typing-extensions==3.10.0.1
uritemplate==3.0.1
urllib3==1.26.6
